feat: removed plausible and updated to latest dependencies
