+++ frontend/.env.example
@@ -23,7 +23,3 @@ CURRENT_BRANCH=dev
 
 
 # Analytics
-
-# Not needed for dev setups but nice for production analytics via plausbile
-PLAUSIBLE_DOMAIN=local.coregame.sh
-PLAUSIBLE_CUSTOM_DOMAIN=https://plausible.coregame.sh
