+++ frontend/app/about/aboutPage.tsx
@@ -1,11 +1,10 @@
 "use client";
 
 import { motion } from "framer-motion";
-import { LucideGithub, LucideLinkedin } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
 import { title } from "@/components/primitives";
-import { CoreLogoWhite } from "@/components/social";
+import { CoreLogoWhite, GithubIcon, LinkedinIcon } from "@/components/icons";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
 
@@ -242,10 +241,10 @@ export default function AboutPageClient() {
                       >
                         {m.linkType === "github"
                           ? (
-                              <LucideGithub />
+                              <GithubIcon />
                             )
                           : (
-                              <LucideLinkedin />
+                              <LinkedinIcon />
                             )}
                       </a>
                     </div>
