+++ frontend/eslint.config.mjs
@@ -1,37 +1,16 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import antfu from "@antfu/eslint-config";
-import tailwind from "eslint-plugin-tailwindcss";
+import { defineConfig, globalIgnores } from 'eslint/config'
+import nextVitals from 'eslint-config-next/core-web-vitals'
 
-const configDir = path.dirname(fileURLToPath(import.meta.url));
-const tailwindEntry = path.join(configDir, "styles", "globals.css");
+const eslintConfig = defineConfig([
+  ...nextVitals,
+  // Override default ignores of eslint-config-next.
+  globalIgnores([
+    // Default ignores of eslint-config-next:
+    '.next/**',
+    'out/**',
+    'build/**',
+    'next-env.d.ts',
+  ]),
+])
 
-export default antfu({
-  type: "app",
-  nextjs: true,
-  formatters: true,
-  typescript: true,
-  stylistic: {
-    indent: 2,
-    semi: true,
-    quotes: "double",
-  },
-  ignores: [".pnpm-store/**", "helm/**"],
-}, {
-  rules: {
-    "no-console": "warn",
-    "no-debugger": "warn",
-    "node/prefer-global/process": "off",
-    "no-alert": "warn",
-  },
-}, ...tailwind.configs["flat/recommended"], {
-  settings: {
-    tailwindcss: {
-      config: tailwindEntry,
-    },
-  },
-  rules: {
-    "tailwindcss/classnames-order": "warn",
-    "tailwindcss/no-custom-classname": "off",
-  },
-});
+export default eslintConfig
