--- frontend/helm/templates/plausible-proxy.yaml
@@ -1,60 +1,0 @@
-{{- if .Values.ingress.enabled -}}
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "frontend.fullname" . }}-plausible
-  labels:
-    {{- include "frontend.labels" . | nindent 4 }}
-spec:
-  type: ExternalName
-  externalName: plausible.coregame.sh
-  ports:
-  - port: 443
-    targetPort: 443
----
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: {{ include "frontend.fullname" . }}-plausible
-  labels:
-    {{- include "frontend.labels" . | nindent 4 }}
-  annotations:
-    nginx.ingress.kubernetes.io/backend-protocol: "HTTPS"
-    nginx.ingress.kubernetes.io/upstream-vhost: "plausible.coregame.sh"
-    nginx.ingress.kubernetes.io/ssl-redirect: "true"
-    nginx.ingress.kubernetes.io/proxy-buffering: "on"
-    nginx.ingress.kubernetes.io/proxy-http-version: "1.1"
-    nginx.ingress.kubernetes.io/use-regex: "true"
-spec:
-  ingressClassName: {{ .Values.ingress.className }}
-  {{- if .Values.ingress.tls }}
-  tls:
-    {{- range .Values.ingress.tls }}
-    - hosts:
-        {{- range .hosts }}
-        - {{ . | quote }}
-        {{- end }}
-      secretName: {{ .secretName }}
-    {{- end }}
-  {{- end }}
-  rules:
-    {{- range .Values.ingress.hosts }}
-    - host: {{ .host | quote }}
-      http:
-        paths:
-        - path: /js/.*\.js
-          pathType: ImplementationSpecific
-          backend:
-            service:
-              name: {{ include "frontend.fullname" $ }}-plausible
-              port:
-                number: 443
-        - path: /api/event
-          pathType: Prefix
-          backend:
-            service:
-              name: {{ include "frontend.fullname" $ }}-plausible
-              port:
-                number: 443
-    {{- end }}
-{{- end }}
