+++ frontend/helm/values-dev.yaml
@@ -24,9 +24,6 @@ env:
   BACKEND_URL: "http://api-dev" # internal API route
   NEXT_PUBLIC_VISUALIZER_URL: "https://dev.visualizer.coregame.sh"
   NEXT_PUBLIC_BACKEND_PUBLIC_URL: "https://dev.api.coregame.sh" # public API route accessible from outside the cluster
-  # Plausible analytics domain defaults for dev
-  PLAUSIBLE_DOMAIN: dev.coregame.sh
-  PLAUSIBLE_CUSTOM_DOMAIN: "https://plausible.coregame.sh"
   AUTH_COOKIE_NAME: "token_dev"
 
 ingress:
