+++ frontend/package.json
@@ -12,47 +12,46 @@
     "lint": "eslint . --fix"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.2.2",
-    "@internationalized/date": "^3.11.0",
-    "@radix-ui/react-accordion": "^1.2.12",
-    "@radix-ui/react-avatar": "^1.1.11",
-    "@radix-ui/react-checkbox": "^1.3.3",
-    "@radix-ui/react-dialog": "^1.1.15",
-    "@radix-ui/react-dropdown-menu": "^2.1.16",
-    "@radix-ui/react-label": "^2.1.8",
-    "@radix-ui/react-navigation-menu": "^1.2.14",
-    "@radix-ui/react-popover": "^1.1.15",
-    "@radix-ui/react-select": "^2.2.6",
-    "@radix-ui/react-slot": "^1.2.4",
-    "@radix-ui/react-switch": "^1.2.6",
-    "@radix-ui/react-tabs": "^1.1.13",
-    "@radix-ui/react-tooltip": "^1.2.8",
-    "@react-aria/visually-hidden": "^3.8.30",
-    "@tanstack/react-form": "^1.28.3",
-    "@tanstack/react-query": "^5.90.21",
+    "@hookform/resolvers": "^5.4.0",
+    "@internationalized/date": "^3.12.2",
+    "@radix-ui/react-accordion": "^1.2.14",
+    "@radix-ui/react-avatar": "^1.2.0",
+    "@radix-ui/react-checkbox": "^1.3.5",
+    "@radix-ui/react-dialog": "^1.1.17",
+    "@radix-ui/react-dropdown-menu": "^2.1.18",
+    "@radix-ui/react-label": "^2.1.10",
+    "@radix-ui/react-navigation-menu": "^1.2.16",
+    "@radix-ui/react-popover": "^1.1.17",
+    "@radix-ui/react-select": "^2.3.1",
+    "@radix-ui/react-slot": "^1.3.0",
+    "@radix-ui/react-switch": "^1.3.1",
+    "@radix-ui/react-tabs": "^1.1.15",
+    "@radix-ui/react-tooltip": "^1.2.10",
+    "@react-aria/visually-hidden": "^3.9.1",
+    "@tanstack/react-form": "^1.33.0",
+    "@tanstack/react-query": "^5.101.2",
     "@tanstack/react-table": "^8.21.3",
-    "axios": "^1.15.0",
+    "axios": "^1.18.1",
     "chart.js": "^4.5.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "date-fns": "^4.1.0",
-    "framer-motion": "^12.34.1",
+    "date-fns": "^4.4.0",
+    "framer-motion": "^12.42.0",
     "gray-matter": "^4.0.3",
-    "lucide-react": "^0.574.0",
-    "next": "^16.2.4",
-    "next-auth": "^4.24.13",
-    "next-plausible": "^3.12.5",
+    "lucide-react": "^1.21.0",
+    "next": "^16.2.9",
+    "next-auth": "^4.24.14",
     "next-themes": "^0.4.6",
-    "react": "^19.2.4",
+    "react": "^19.2.7",
     "react-chartjs-2": "^5.3.1",
-    "react-day-picker": "^9.13.2",
-    "react-dom": "^19.2.4",
-    "react-hook-form": "^7.71.1",
+    "react-day-picker": "^10.0.1",
+    "react-dom": "^19.2.7",
+    "react-hook-form": "^7.80.0",
     "reactflow": "^11.11.4",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-code-group": "^0.3.1",
     "rehype-parse": "^9.0.1",
-    "rehype-pretty-code": "^0.14.1",
+    "rehype-pretty-code": "^0.14.3",
     "rehype-raw": "^7.0.0",
     "rehype-sanitize": "^6.0.0",
     "rehype-slug": "^6.0.0",
@@ -62,34 +61,35 @@
     "remark-html": "^16.0.1",
     "remark-parse": "^11.0.0",
     "remark-rehype": "^11.1.2",
-    "shiki": "^3.22.0",
+    "shiki": "^4.3.0",
+    "simple-icons": "^16.24.0",
     "sonner": "^2.0.7",
     "strip-markdown": "^6.0.0",
-    "tailwind-merge": "^3.4.1",
+    "tailwind-merge": "^3.6.0",
     "tailwind-variants": "^3.2.2",
     "unified": "^11.0.5",
-    "zod": "^4.3.6"
+    "zod": "^4.4.3"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^7.4.3",
-    "@eslint-react/eslint-plugin": "^2.13.0",
-    "@next/eslint-plugin-next": "^16.1.6",
-    "@tailwindcss/postcss": "^4.1.18",
-    "@tailwindcss/typography": "^0.5.19",
-    "@types/node": "^25.2.3",
-    "@types/react": "^19.2.14",
+    "@eslint-react/eslint-plugin": "^5.10.0",
+    "@next/eslint-plugin-next": "^16.2.9",
+    "@tailwindcss/postcss": "^4.3.1",
+    "@tailwindcss/typography": "^0.5.20",
+    "@types/node": "^26.0.1",
+    "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
-    "eslint": "^9.39.2",
-    "eslint-plugin-format": "^1.4.0",
-    "eslint-plugin-react-hooks": "^7.0.1",
-    "eslint-plugin-react-refresh": "^0.5.0",
-    "eslint-plugin-tailwindcss": "4.0.0-beta.0",
-    "lint-staged": "^16.2.7",
-    "postcss": "^8.5.6",
-    "prettier": "^3.8.1",
-    "tailwindcss": "^4.1.18",
+    "eslint": "^9.39.4",
+    "eslint-config-next": "^16.2.9",
+    "eslint-plugin-format": "^2.0.1",
+    "eslint-plugin-react-hooks": "^7.1.1",
+    "eslint-plugin-react-refresh": "^0.5.3",
+    "eslint-plugin-tailwindcss": "4.0.4",
+    "lint-staged": "^17.0.8",
+    "postcss": "^8.5.15",
+    "prettier": "^3.9.1",
+    "tailwindcss": "^4.3.1",
     "tw-animate-css": "^1.4.0",
-    "typescript": "^5.9.3"
+    "typescript": "^6.0.3"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx,json,css,scss,md}": [
