+++ frontend/scripts/clone-wiki-branches.js
@@ -1,4 +1,4 @@
-/* eslint-disable no-console */
+ 
 const { execSync } = require("node:child_process");
 const fs = require("node:fs");
 const path = require("node:path");
